nfc: rework mfu poller
