+++ lib/nfc/protocols/mf_ultralight/mf_ultralight_poller.h
@@ -36,70 +36,7 @@ typedef struct {
     MfUltralightPollerEventData* data;
 } MfUltralightPollerEvent;
 
-typedef enum {
-    MfUltralightPollerCommandContinue = NfcaPollerCommandContinue,
-    MfUltralightPollerCommandReset = NfcaPollerCommandReset,
-    MfUltralightPollerCommandStop = NfcaPollerCommandStop,
-} MfUltralightPollerCommand;
-
-typedef MfUltralightPollerCommand (
-    *MfUltralightPollerCallback)(MfUltralightPollerEvent event, void* context);
-
-MfUltralightPoller* mf_ultralight_poller_alloc(NfcaPoller* nfca_poller);
-
-void mf_ultralight_poller_free(MfUltralightPoller* instance);
-
 const MfUltralightData* mf_ultralight_poller_get_data(MfUltralightPoller* instance);
-
-MfUltralightError mf_ultralight_poller_start(
-    MfUltralightPoller* instance,
-    NfcaPollerEventCallback callback,
-    void* context);
-
-MfUltralightError mf_ultralight_poller_read(
-    MfUltralightPoller* instance,
-    MfUltralightPollerCallback callback,
-    void* context);
-
-MfUltralightError mf_ultralight_poller_reset(MfUltralightPoller* instance);
-
-MfUltralightError mf_ultralight_poller_stop(MfUltralightPoller* instance);
-
-// Sync API
-
-MfUltralightError mf_ultralight_poller_read_page(
-    MfUltralightPoller* instance,
-    uint16_t page,
-    MfUltralightPage* data);
-
-MfUltralightError mf_ultralight_poller_read_page_new(
-    NfcPollerManager* poller_manager,
-    uint16_t page,
-    MfUltralightPage* data);
-
-MfUltralightError mf_ultralight_poller_write_page(
-    MfUltralightPoller* instance,
-    uint16_t page,
-    MfUltralightPage* data);
-
-MfUltralightError
-    mf_ultralight_poller_read_version(MfUltralightPoller* instance, MfUltralightVersion* data);
-
-MfUltralightError
-    mf_ultralight_poller_read_signature(MfUltralightPoller* instance, MfUltralightSignature* data);
-
-MfUltralightError mf_ultralight_poller_read_counter(
-    MfUltralightPoller* instance,
-    uint8_t counter_num,
-    MfUltralightCounter* data);
-
-MfUltralightError mf_ultralight_poller_read_tearing_flag(
-    MfUltralightPoller* instance,
-    uint8_t flag_num,
-    MfUltralightTearingFlag* data);
-
-MfUltralightError
-    mf_ultralight_poller_read_card(MfUltralightPoller* instance, MfUltralightData* data);
 
 #ifdef __cplusplus
 }
